ChessGame/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
click==8.1.8
Django==5.1.7
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.1
wsproto==1.2.0
